pyproject.toml
[project]
name = "chatdoctor"
version = "2.0.0"
description = "ChatDoctor - A Medical AI Chatbot fine-tuned on LLaMA using medical domain knowledge"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "ChatDoctor Team"},
]
keywords = ["medical", "ai", "chatbot", "llm", "healthcare", "llama", "fine-tuning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.38.0",
    "peft>=0.8.0",
    "bitsandbytes>=0.42.0",
    "accelerate>=0.25.0",
    "datasets>=2.16.0",
    "huggingface-hub>=0.20.0",
    "sentencepiece",
    "fire",
    "tqdm",
    "pandas",
    "numpy>=1.24.0",
    "scipy",
    "pyyaml",
    "rich",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
]
eval = [
    "rouge_score",
    "scikit-learn",
    "nltk",
]
rag = [
    "llama-index>=0.10.0",
    "wikipedia",
]
wandb = [
    "wandb",
]
openai = [
    "openai>=1.0.0",
]
all = [
    "chatdoctor[dev,eval,rag,wandb,openai]",
]

[project.scripts]
chatdoctor = "chatdoctor.cli:main"
chatdoctor-train = "chatdoctor.training.train_lora:main"
chatdoctor-chat = "chatdoctor.inference.chat:main"

[project.urls]
Homepage = "https://github.com/Kent0n-Li/ChatDoctor"
Documentation = "https://github.com/Kent0n-Li/ChatDoctor"
Repository = "https://github.com/Kent0n-Li/ChatDoctor"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["chatdoctor*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
